% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{All.filtered}
\alias{All.filtered}
\title{Bacterial growth curves in 4 different media}
\format{A data.frame with the following columns
\describe{
  \item{Row}{Row letter in the 96-well plate wee bacteria was grown}
  \item{Column}{Column from the 96-well plate were bacteria was grown}
  \item{OD600}{Bacterial abundance measurment in optical density at 600nm}
  \item{Rep}{Replicate number}
  \item{hrs}{Time-point of measurment in hours}
  \item{Strain}{Strain ID}
  \item{Condition}{Media used to grow the bacteria. minusP is Johnson
  media with 0 phosphate added, plusP is Johnson media with 1mM phosphate
  added. For minus2plusP, Arabidopsis seeds were germinated in Johnson
  media with zero phosphate, and then transferred for 24 hours into
  Johnson media with 1mM phosphate. After 24h the media was filter-sterilized
  and bacteria were grown on it. For plus2minusP Arabidopsis seeds were
  germinated in Johnson media with 1mM phosphate, and then transferred
  for 24 hours into Johnson media with no phosphate. After 24h the
  media was filter-sterilized}
}}
\source{
Dangl Lab
}
\usage{
All.filtered
}
\description{
A data.frame containing the 72-hour long of 441 strains
grown in 4 conditions. Growth curves have been filtered
to remove strains that didn't grow by clustering and removing
those strains that looked more similar to sample blanks
}
\keyword{datasets}
