% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_effects.r
\name{block_effects}
\alias{block_effects}
\title{Estimate block main effects}
\usage{
block_effects(Dat, var.name, cond1, cond2, keep.vars = c("Plate",
  "Experiment"), create.design = TRUE)
}
\description{
Estimate block main effects
}
