% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist2df.r
\name{dist2df}
\alias{dist2df}
\title{Distance to data.frame}
\usage{
dist2df(inDist)
}
\arguments{
\item{inDist}{An object of class "dist". See \code{\link{dist}} for
more info}
}
\description{
Convets an object of class "dist" into an
obejct of class "data.frame" where each row is
a pairwise distance
}
\seealso{
\code{\link{dist}}
}
\author{
Sur Herrera Paredes
}
\keyword{utilities}
