% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.r
\name{homogenize_taxa}
\alias{homogenize_taxa}
\title{homogenize taxa in abundance matrix}
\usage{
homogenize_taxa(Tab, taxa.list)
}
\arguments{
\item{Tab}{Abundance matrix}

\item{taxa.list}{character vector of taxa

Internal}
}
\description{
homogenize taxa in abundance matrix
}
