% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/independent_lm_tests.r
\name{independent_lm_tests}
\alias{independent_lm_tests}
\title{Test all combinations of levels}
\usage{
independent_lm_tests(Dat, f1, strain, StartP.levels = c("minusP", "plusP"),
  EndP.levels = c("100uM", "30uM"))
}
\arguments{
\item{Dat}{data.frame from one bacteria}

\item{f1}{formula for test}

\item{strain}{strain name to test}

\item{StartP.levels}{new names for StartP levels}

\item{EndP.lvels}{new names for endP levels.}
}
\value{
A list
}
\description{
Internal function.
}
\author{
Sur Herrera Paredes
}
