% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_phen_and_abun.r
\name{merge_phen_and_abun}
\alias{merge_phen_and_abun}
\title{Merge phenotype data with relative abundances of blocks}
\usage{
merge_phen_and_abun(Phen, abun, columns, varnames = c("Bacteria",
  "Experiment", "StartP", "EndP"))
}
\arguments{
\item{Phen}{A data frame with phenotypic information}

\item{abun}{A data frame with abundance information}

\item{varnames}{Vector of charachter strings indicating which
variables to use to combine datasets. Rows where the exact same
combination of values for all variables occur are considered identical.}
}
\description{
Merge two data frames containing plant phenotypic data
and bacterial relative abundances
}
\details{
The function takes a list of variables that when combined must
uniqueley identify each row of both data frames. Those variable smust exist
in both data frames.

It then constructs an index that it uses to match the columns. Abundance
data can be missing for some phenotypic observations, in which case
it will be replaced by zero.
}
\author{
Sur Herrera Paredes
}
\keyword{colonization}
\keyword{syncom}
\keyword{utils}
