% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacoder_plot_go.r
\name{metacoder_plot_go}
\alias{metacoder_plot_go}
\title{Plot GO from edgeR}
\usage{
metacoder_plot_go(dat, output_folder = "./", output_format = "svg",
  min_fdr = 0.01, type = GO.db::GOBPPARENTS, prefix = "go",
  n.supertaxa = 9, num.changed = 3)
}
\description{
Plot GO from edgeR
}
\author{
from metacoder paper
}
