% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtain_block_abundances.r
\name{metadata_from_rowname}
\alias{metadata_from_rowname}
\title{Extract metadata from rownames}
\usage{
metadata_from_rowname(Dat, varnames = c("Bacteria", "Replicate", "Experiment",
  "Pre.Pi", "Pos.Pi"), sep = "_")
}
\description{
Extract metadata from rownames
}
