% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_tax_and_plot.r
\name{mgsub}
\alias{mgsub}
\title{mgsub}
\usage{
mgsub(pattern, replacement, x, ...)
}
\description{
Internal
}
\references{
from: http://stackoverflow.com/questions/15253954/replace-multiple-arguments-with-gsub
}
