% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtain_block_abundances.r
\name{obtain_block_abundances}
\alias{obtain_block_abundances}
\title{Obtain block abundances}
\usage{
obtain_block_abundances(Dat, varnames = c("Bacteria", "Replicate",
  "Experiment", "Pre.Pi", "Pos.Pi"), taxa.group = "Block", sep = "_",
  taxa2rm = "contaminant")
}
\description{
Takes a Dataset object, pools samples and taxa according to
specified variables, and returns  adata frame that contains
the average abundance by taxa group per group of samples
}
\author{
Sur Herrera Paredes
}
