% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_tax_and_plot.r
\name{parse_tax_and_plot}
\alias{parse_tax_and_plot}
\title{Metacoder plot}
\usage{
parse_tax_and_plot(file, col, output_file = output_file,
  n.supertaxa = n.supertaxa, num.changed = num.changed, min_fdr = min_fdr)
}
\description{
Requires you to load metacoder package, otherwise names not print
}
\author{
From metacoder paper
}
