% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_combined_effects.r
\name{plot_combined_effects}
\alias{plot_combined_effects}
\title{Plot heatmap of all effects}
\usage{
plot_combined_effects(dat, scale = TRUE, cluster = TRUE,
  standardize = FALSE)
}
\description{
Plot heatmap of all effects
}
