% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_res.r
\name{plot_res}
\alias{plot_res}
\title{Plot binary association results}
\usage{
plot_res(test, qval_thres = 0.05)
}
\arguments{
\item{test}{result from test all strains}

\item{qval_trhes}{q-value threshold}
}
\value{
A ggplot2 object
}
\description{
Plot binary association results
}
\author{
Sur Herrera Paredes
}
