% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_set.r
\name{plot_set}
\alias{plot_set}
\title{Plot the expression of a set of genes}
\usage{
plot_set(Dat, gene.set, groups)
}
\arguments{
\item{Dat}{Dataset. Should already be normalized}

\item{gene.set}{character vector of gene IDs to plot}

\item{groups}{character vector of column names of groups
to plot}
}
\value{
A dataset
}
\description{
Boxplot with dots of set of normalized RPKM values
based on a grouping factor.
}
\author{
Sur Herrera Paredes
}
