% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotgg_heatmap_syncom_efffects.r
\name{plotgg_heatmap_syncom_effects}
\alias{plotgg_heatmap_syncom_effects}
\title{Plot syncom effects}
\usage{
plotgg_heatmap_syncom_effects(Res, cond1, cond2)
}
\description{
DEPRECATED
}
