% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_smooth_func.r
\name{stat_smooth_func}
\alias{stat_smooth_func}
\title{Stat_smooth for coef of determination}
\usage{
stat_smooth_func(mapping = NULL, data = NULL, geom = "smooth",
  position = "identity", ..., method = "auto", formula = y ~ x,
  se = TRUE, n = 80, span = 0.75, fullrange = FALSE, level = 0.95,
  method.args = list(), na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, xpos = NULL, ypos = NULL)
}
\description{
https://gist.github.com/kdauria/524eade46135f6348140
}
