% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_class.r
\name{term_class}
\alias{term_class}
\title{From metacoder}
\usage{
term_class(x, current = x, all_paths = TRUE, type = GO.db::GOCCPARENTS,
  verbose = TRUE, valid_relationships = c("is_a"))
}
\description{
From metacoder
}
\author{
From Metacoder paper
}
