% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_all_strains.r
\name{test_all_strains}
\alias{test_all_strains}
\title{Test all strains for change in phosphate accumulation}
\usage{
test_all_strains(Strains, treatment_col, treatment_control = "No Bacteria",
  Master, plot = FALSE, StartP.levels = c("minusP", "plusP"),
  EndP.levels = c("100uM", "30uM"))
}
\arguments{
\item{Strains}{vector of strain names}

\item{treatment_col}{Column name with variable to test}

\item{Master}{data.frame with all data}

\item{plot}{logical produce plot?}

\item{StartP.levels}{New neames for StartP levels. It must be in the
same order as the levels in Master}

\item{EndP.levels}{New names for EndP levels. Must be in the same order
as levels in Master.}

\item{treament_control}{level to use as reference for anova}
}
\value{
A list.
}
\description{
Performs anova in every strain and returns log-fold
change versus no bacteria. Performs anova independently
in eveyr combination of StartP and EndP levels
}
\author{
Sur Herrera Paredes
}
