% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_auc.R
\name{test_auc}
\alias{test_auc}
\title{Test AUC}
\usage{
test_auc(Dat, thres = 0.05)
}
\arguments{
\item{Dat}{A data.frame with the same columns as \link{All.filtered}}

\item{thres}{q-value threshold for significance}
}
\description{
Tests area under the curve for a set of
strains
}
\details{
Assumes same number of timepoints per growth curve
}
