% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_single_block_phenotype.r
\name{test_single_block_phenptype}
\alias{test_single_block_phenptype}
\title{Test the effect of single block on phenotype}
\usage{
test_single_block_phenptype(Phen, abun, phenotype, variables = c("Bacteria",
  "Experiment", "StartP", "EndP"), confounders = c("Experiment"),
  use_abun = TRUE)
}
\description{
If abundance is not used, it uses simply 1/0 for
presence/absence
}
