% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_single_community_phenotype.r
\name{test_single_community_phenotype}
\alias{test_single_community_phenotype}
\title{Test the effect of all single communities}
\usage{
test_single_community_phenotype(Dat, dir, var.name, bacteria.col = "Bacteria",
  ref.level = "none", plot = FALSE, f1.extra = "+ Experiment + Plate")
}
\description{
Test the effect of all single communities
}
