% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/independent_lm_tests.r
\name{independent_lm_tests}
\alias{independent_lm_tests}
\title{Test all combinations of levels}
\usage{
independent_lm_tests(Dat, f1, strain, StartP.levels = c("minusP", "plusP"),
  EndP.levels = c("100uM", "30uM"))
}
\arguments{
\item{Dat}{A data.frame containing the data of one bacterial treatment.
Unsupported behaviour would occur if more than one bacterial treatment is
included. The data.frame must contain columns: 'Bacteria', 'StartP' and
'EndP'. The 'Bacteria' column must be either a factor or a character vector
containing only values 'No Bacteria' and '+Bacteria'}

\item{f1}{formula for test. It will be passed to the lm function}

\item{strain}{A character string, indicating strain name.}

\item{StartP.levels}{new names for StartP levels}

\item{EndP.lvels}{new names for EndP levels.}
}
\value{
A list of data frames. Each data.frame contains a column called 'Strain'
which indicates the strain ID. An also contains m more columns
(m = length(StartP.levels)*lenght(EndP.levels)). Each element of the list stores the
either the coefficiennt for the bacterial treatment (Res), its p-value (Pval), or
its standard error (SE).
}
\description{
Internal function.
}
\details{
This functions takes the data from one bacterial treatment makes
an independent linear model for all observation on each combination 
of levels between two variables.
}
\seealso{
\code{\link{lm}}
}
\author{
Sur Herrera Paredes
}
\keyword{binP}
\keyword{syncom}
\keyword{utilities}
