% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_phen_and_abun.r
\name{merge_phen_and_abun}
\alias{merge_phen_and_abun}
\title{Merge phenotype data with relative abundances of blocks}
\usage{
merge_phen_and_abun(Phen, abun, columns, varnames = c("Bacteria",
  "Experiment", "StartP", "EndP"))
}
\description{
Merge phenotype data with relative abundances of blocks
}
